refactored into pres and container component and hooks
